feat(player): extract AudiobookPlayerViewModel and related types to feature:player KMP module
Closes #868

Move AudiobookPlayerViewModel, AudiobookReconciliationCoordinator, FollowLoopOrchestrator,
AudiobookResumeResolver, AudiobookHandoffState, AudiobookPlayerEvent, AudiobookPlayerUiState,
AudiobookProgressUtils, NowPlaying, NowPlayingStore, SleepTimerMode, and ReadaloudHandoff to a
new `feature:player` KMP module (commonMain + jvmMain). Introduce AudioPlayerInterface so
AudiobookController stays Android-only (in :app) but implements a pure-Kotlin interface the
shared ViewModel can depend on.

Key structural changes:
- feature:player: new KMP module with jvm() + iosArm64 + iosSimulatorArm64 targets
- feature:reader: promote AbsTargetResolver, PositionSaveCoordinator, ProgressFlushScope to
  commonMain; promote ReaderSync + ReaderSyncFactory to jvmMain; add ReaderSyncInterfaces.kt
  with commonMain interface types (ReaderSyncCoordinatorInterface, AudiobookFollowInterface,
  ReaderSyncFactoryInterface, AudioLedCycleResult)
- core:domain: add CrossEpubIndexBuildTrigger and PlaylistsRepository (moved from core:data
  which has no JVM target); core:data keeps typealias bridges for backward compatibility
- AudiobookPlayerViewModel: replace SavedStateHandle constructor param with direct nav arg
  params (navItemId, navPlaylistId, navPlaylistLibraryId, navStartAtSec)
- Koin DI: AudiobookController binds AudioPlayerInterface::class; ReadaloudController binds
  ReadaloudHandoff::class; ReaderSyncFactory binds ReaderSyncFactoryInterface::class
- AbsProgressPeerTest moved from app:test to feature:reader:jvmTest where internal peer
  classes are accessible; test imports updated to kotlin.test and JSON string literals

All JVM tests pass; KoinModuleVerificationTest green; checkNoAndroidImports,
checkNoServerReferences, checkRiffleLogTags, checkTestGuardrails all pass.
