+++ app/build.gradle.kts
@@ -127,6 +127,7 @@ dependencies {
     implementation(project(":core:domain"))
     implementation(project(":feature:library"))
     implementation(project(":feature:reader"))
+    implementation(project(":feature:player"))
     implementation(project(":feature:source"))
     implementation(project(":core:sources"))
     implementation(project(":core:sync"))
