+++ app/src/main/kotlin/com/riffle/app/feature/audiobook/AudiobookPlayerViewModel.kt
@@ -1,961 +1,6 @@
 package com.riffle.app.feature.audiobook
 
-import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.riffle.core.models.AudioIdentity
-import com.riffle.core.domain.AudioIdentityResolver
-import com.riffle.core.domain.AudioPlaybackPreferencesStore
-import com.riffle.core.models.AudiobookBookmark
-import com.riffle.core.domain.ContentCacheAccessStore
-import com.riffle.core.domain.ContentCacheArtifactKind
-import com.riffle.core.domain.ContentCacheKey
-import com.riffle.core.domain.ListeningPreferencesStore
-import com.riffle.core.domain.AudiobookBookmarkStore
-import com.riffle.core.domain.AudiobookChapter
-import com.riffle.core.domain.AudiobookCacheRepository
-import com.riffle.core.domain.AudiobookRepository
-import com.riffle.core.domain.AudiobookTimeline
-import com.riffle.core.domain.BookmarkTitleBuilder
-import com.riffle.core.common.Clock
-import com.riffle.core.domain.LibraryObserver
-import com.riffle.core.domain.usecase.UpdateReadingProgress
-import com.riffle.core.domain.SourceRepository
-import com.riffle.core.domain.TokenStorage
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.asSharedFlow
-import kotlinx.coroutines.flow.filter
-import kotlinx.coroutines.flow.filterNotNull
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.CompletableDeferred
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withTimeoutOrNull
-import com.riffle.core.logging.LogChannel
-import com.riffle.core.logging.Logger
-import kotlin.math.roundToInt
-
-/**
- * The book-absolute resume position. When NO position was tracked at all ([hadTrackedPosition] false —
- * no local audiobook row and no server record, e.g. offline with only a downloaded bundle) fall back
- * to the item's library [readingProgressFraction] mapped to seconds. This resumes near where the rest
- * of the app shows progress instead of restarting at 0, and (crucially) seeds the resume floor so the
- * close/follow persist guard never writes a ~0 over that progress — the offline-erase bug.
- *
- * Gating on [hadTrackedPosition] — not on `reconciledSec > 0` — is deliberate: a genuinely-tracked
- * position of exactly 0 (a server record or local row that says "at the start", with a real
- * timestamp) must be honoured, not replaced by the progress fallback. Online behaviour is unchanged
- * because a server record always counts as a tracked position.
- */
-internal fun audiobookResumeSec(
-    reconciledSec: Double,
-    hadTrackedPosition: Boolean,
-    readingProgressFraction: Float,
-    durationSec: Double,
-): Double =
-    if (hadTrackedPosition || readingProgressFraction <= 0f || durationSec <= 0.0) reconciledSec
-    else (readingProgressFraction * durationSec).coerceIn(0.0, durationSec)
-
-/** How close to the end a resume position must sit to count the book as finished (and restart at 0). */
-internal const val AUDIOBOOK_FINISHED_EPS_SEC: Double = 1.0
-
-/**
- * The position to actually start a normally-opened audiobook at. A resume that lands at (or within
- * [AUDIOBOOK_FINISHED_EPS_SEC] of) the end means the book was finished: seeding the player there puts
- * it at the end of the last track, where ExoPlayer is `STATE_ENDED` and `play()` is a no-op — the
- * player sits silent with the seek bar pinned at the end. Reopening a finished book is a replay
- * intent, so restart from the beginning instead. A position anywhere short of the end is honoured
- * as-is. No-op when the duration is unknown (we can't tell what "the end" is).
- */
-internal fun audiobookStartSec(resumeSec: Double, durationSec: Double): Double =
-    if (durationSec > 0.0 && resumeSec >= durationSec - AUDIOBOOK_FINISHED_EPS_SEC) 0.0 else resumeSec
-
-/**
- * The one-line facts shown beneath the cover on the landscape player. Duration is omitted when
- * unknown; at most two genres are listed to keep it tidy. Null when there's nothing to show.
- */
-internal data class CompactDurationLabelTemplates(
-    val minutes: String = "%1\$dm",
-    val hours: String = "%1\$dh",
-    val hoursMinutes: String = "%1\$dh %2\$dm",
-)
-
-internal fun formatCompactDuration(
-    durationSec: Double,
-    templates: CompactDurationLabelTemplates = CompactDurationLabelTemplates(),
-    roundToNearestMinute: Boolean = false,
-): String {
-    val totalMinutes = if (roundToNearestMinute) {
-        (durationSec / 60.0).roundToInt().coerceAtLeast(0)
-    } else {
-        (durationSec.toLong().coerceAtLeast(0) / 60).toInt()
-    }
-    val hours = totalMinutes / 60
-    val minutes = totalMinutes % 60
-    return when {
-        hours > 0 && minutes > 0 -> templates.hoursMinutes.format(hours, minutes)
-        hours > 0 -> templates.hours.format(hours)
-        else -> templates.minutes.format(minutes)
-    }
-}
-
-internal fun buildAudiobookFacts(
-    durationSec: Double,
-    genres: List<String>,
-    audiobookLabel: String = "Audiobook",
-    durationLabels: CompactDurationLabelTemplates = CompactDurationLabelTemplates(),
-): String? {
-    val parts = buildList {
-        add(audiobookLabel)
-        if (durationSec > 0.0) {
-            add(formatCompactDuration(durationSec, durationLabels))
-        }
-        genres.take(2).forEach { add(it) }
-    }
-    // The medium label alone has no real facts, so it is not worth a line.
-    return if (parts.size > 1) parts.joinToString(" · ") else null
-}
-
-/** One-shot UI events emitted by the [AudiobookPlayerViewModel] — collected by the screen. */
-sealed interface AudiobookPlayerEvent {
-    /** Playback finished naturally (last track ended); the screen should close the player. */
-    data object Finished : AudiobookPlayerEvent
-
-    /**
-     * Playback finished AND the player was opened inside a playlist context that has a next
-     * item. The screen navigates to the next item's audiobook player (preserving the same
-     * playlist context) so playback auto-advances through the playlist.
-     */
-    data class PlaylistAdvance(val nextItemId: String) : AudiobookPlayerEvent
-}
-
-/** UI state for the full-screen [Audiobook Player] (ADR 0035). */
-data class AudiobookPlayerUiState(
-    val loading: Boolean = true,
-    val failed: Boolean = false,
-    val title: String = "",
-    val author: String = "",
-    val publishedYear: String? = null,
-    val coverUrl: String? = null,
-    val authToken: String = "",
-    val isPlaying: Boolean = false,
-    val speed: Float = 1f,
-    val positionSec: Double = 0.0,
-    val durationSec: Double = 0.0,
-    // Book-absolute position up to which audio has buffered ahead of [positionSec]. Rendered as a
-    // lighter fill on the seek bar, YouTube/Spotify-style.
-    val bufferedPositionSec: Double = 0.0,
-    val currentChapterTitle: String? = null,
-    val chapterStartsSec: List<Double> = emptyList(),
-    // The full chapter list + the index of the chapter the playhead is in, for the Chapters sheet.
-    val chapters: List<AudiobookChapter> = emptyList(),
-    val currentChapterIndex: Int = -1,
-    val canPreviousChapter: Boolean = false,
-    val canNextChapter: Boolean = false,
-    // Book details for the landscape two-column player: raw facts plus the blurb (ADR 0035).
-    val genres: List<String> = emptyList(),
-    val facts: String? = null,
-    val description: String? = null,
-    // The linked readaloud EBOOK item id, when this title has one (split-library ebook, or this same
-    // item if it's a combined ebook+audio). Non-null enables swipe-down → switch to the readaloud
-    // reader; null means there's no readaloud to switch to, so no swipe-down.
-    val readaloudEbookItemId: String? = null,
-    // User bookmarks for this audiobook, observed live from the store (ordered by position, earliest first).
-    val bookmarks: List<AudiobookBookmark> = emptyList(),
-    // True only when this item has unsynced (dirty) bookmarks AND the device is offline, so the
-    // Bookmarks sheet can show a quiet "Offline — bookmarks will sync" note. Sync is otherwise silent.
-    val bookmarksOffline: Boolean = false,
-    val sleepTimer: SleepTimerMode = SleepTimerMode.None,
-    val skipIntervalSeconds: Int = 30,
-    val rewindIntervalSeconds: Int = 15,
-)
-
-class AudiobookPlayerViewModel constructor(
-    savedStateHandle: SavedStateHandle,
-    private val audiobookRepository: AudiobookRepository,
-    private val audiobookDownloadRepository: com.riffle.core.domain.AudiobookDownloadRepository,
-    private val audiobookCacheRepository: AudiobookCacheRepository,
-    private val bundleAudiobookSource: com.riffle.core.domain.BundleAudiobookSource,
-    private val libraryObserver: LibraryObserver,
-    private val updateReadingProgressUseCase: UpdateReadingProgress,
-    private val sourceRepository: SourceRepository,
-    private val tokenStorage: TokenStorage,
-    private val controller: AudiobookController,
-    private val readaloudController: com.riffle.app.feature.reader.readaloud.ReadaloudController,
-    private val audioPlaybackPreferencesStore: AudioPlaybackPreferencesStore,
-    private val listeningPreferencesStore: ListeningPreferencesStore,
-    private val audioIdentityResolver: AudioIdentityResolver,
-    private val readaloudLinkRepository: com.riffle.core.domain.ReadaloudLinkRepository,
-    private val readaloudAudioRepository: com.riffle.core.domain.ReadaloudAudioRepository,
-    private val nowPlayingStore: com.riffle.app.playback.NowPlayingStore,
-    private val audiobookPositionStore: com.riffle.core.domain.AudiobookPositionStore,
-    // While the player is open it drives the book's reconciliation; the durable sweep skips it so it
-    // can't absorb a cross-device server-win the player hasn't seeked to (ADR 0036).
-    private val openReconcileTargets: com.riffle.core.sync.OpenReconcileTargets,
-    private val progressFlushScope: com.riffle.app.feature.reader.ProgressFlushScope,
-    private val bookmarkStore: AudiobookBookmarkStore,
-    private val connectivityObserver: com.riffle.core.domain.ConnectivityObserver,
-    private val audiobookHandoffState: AudiobookHandoffState,
-    private val followLoopOrchestrator: FollowLoopOrchestrator,
-    private val resumeResolver: AudiobookResumeResolver,
-    private val reconciliationCoordinator: AudiobookReconciliationCoordinator,
-    private val clock: Clock,
-    private val logger: Logger,
-    private val playlistsRepository: com.riffle.core.data.PlaylistsRepository,
-    private val contentCacheAccessStore: ContentCacheAccessStore,
-    private val progressSweep: com.riffle.core.sync.ProgressSweep,
-) : ViewModel() {
-
-    private val itemId: String = savedStateHandle.get<String>("itemId") ?: ""
-
-    /** Optional playlist context — set when the player was opened from [PlaylistDetailScreen]'s
-     *  Play button. Empty string / null means "no playlist"; on end-of-book the VM falls back to
-     *  emitting [AudiobookPlayerEvent.Finished] and the screen pops as before. */
-    private val playlistId: String? = savedStateHandle.get<String>("playlistId")?.takeIf { it.isNotEmpty() }
-    private val playlistLibraryId: String? = savedStateHandle.get<String>("libraryId")?.takeIf { it.isNotEmpty() }
-
-    // readaloud→audiobook swipe handoff: the listen position to continue from.
-    // -2 = PREWARM_SENTINEL: pre-warm mode (overlay always-mounted, actual position arrives via
-    //      AudiobookHandoffState when the user swipes up — controller.prepare() is deferred).
-    // -1 = normal resume from saved position.
-    // ≥0 = explicit start second (handoff from the standalone reader path).
-    private val startAtSec: Double = (savedStateHandle.get<Float>("startAtSec") ?: -1f).toDouble()
-
-    // Stored during pre-warm init so activateFromHandoff() can prepare the controller without
-    // re-fetching. Set even in normal-open mode so re-activations (second+ swipe-up) can
-    // re-prepare the controller after releaseForHandoff() clears the spans.
-    private var resolvedSession: com.riffle.core.domain.AudiobookSession? = null
-    // Completed (with the session or null on failure) when init finishes, so activateFromHandoff()
-    // can await rather than drop when a swipe arrives before the ~1–2 s bundle-session fetch is done.
-    private val sessionDeferred = CompletableDeferred<com.riffle.core.domain.AudiobookSession?>()
-    private var resolvedCoverUri: String? = null
-    private var resolvedBookTitle: String? = null
-    private var resolvedInitialSpeed: Float = 1f
-
-    private val meta = MutableStateFlow(
-        AudiobookPlayerUiState(loading = true),
-    )
-    private var timeline: AudiobookTimeline = AudiobookTimeline(0.0)
-    private var sourceId: String = ""
-
-    // The per-book audio-settings key (ADR 0040). The audiobook's own (sourceId, itemId) IS the
-    // canonical key when an audiobook exists, and a linked Readaloud resolves to that same id, so the
-    // speed is shared between them automatically; an audiobook with no bundle gets its own row.
-    private var audioSettingsIdentity: AudioIdentity = AudioIdentity("", itemId)
-    // Debounced-save bookkeeping for the granular speed sheet (mirrors EpubReaderViewModel): the
-    // settled value is written once rather than per intermediate 0.05× step. `pendingSpeed` holds the
-    // not-yet-persisted value so onCleared can flush it.
-    private var pendingSpeed: Float? = null
-    private var speedSaveJob: kotlinx.coroutines.Job? = null
-
-    // The timestamp the local listen position was last genuinely set at; adopted from server stamps
-    // after a write/jump so our own writes never read back as "newer" and re-seek (feedback loop).
-    private var localUpdatedAt: Long = 0L
-
-    /** Adapter that lets [followLoopOrchestrator] read/write this VM's per-book follow state. */
-    private val followContext = object : FollowContext {
-        override fun currentAudioSec(): Double = controller.currentAbsoluteSec()
-        override fun isPlaying(): Boolean = controller.state.value.isPlaying
-        override fun seekTo(positionSec: Double) { controller.seekTo(positionSec) }
-        override val readerSync: com.riffle.app.feature.reader.ReaderSyncCoordinator?
-            get() = reconciliationCoordinator.readerSync
-        override suspend fun tryAttachReaderSync(currentAudioSec: Double): Boolean =
-            attachReaderSync(currentAudioSec, 0L)
-        override fun hasServer(): Boolean = sourceId.isNotEmpty()
-        override fun progressFraction(positionSec: Double): Float =
-            audiobookProgressFraction(positionSec, timeline.durationSec)
-        override suspend fun onHotPathAdvance(positionSec: Double) {
-            positionSaveCoordinator.onChanged(positionSec)
-        }
-        override suspend fun writeSinglePeerFallback(positionSec: Double) {
-            audiobookRepository.saveProgress(sourceId, itemId, positionSec, timeline.durationSec)
-        }
-        override suspend fun writeCloseFlush(positionSec: Double, fraction: Float) {
-            // Persist the trailing second-delta between the last throttled hot-path tick and the
-            // close: onChanged writes the position through savePosition (which also mirrors to the
-            // ebook store and readaloud resume). onClose then updates the library `readingProgress`
-            // float. Prior code dropped the position save here, so audiobook_positions and the
-            // readaloud resume stayed at the last tick and the next reopen resumed a few seconds
-            // behind (#528).
-            positionSaveCoordinator.onChanged(positionSec)
-            positionSaveCoordinator.onClose(fraction)
-        }
-        override var reconciledResumeSec: Double
-            get() = this@AudiobookPlayerViewModel.reconciledResumeSec
-            set(value) { this@AudiobookPlayerViewModel.reconciledResumeSec = value }
-        override var localUpdatedAt: Long
-            get() = this@AudiobookPlayerViewModel.localUpdatedAt
-            set(value) { this@AudiobookPlayerViewModel.localUpdatedAt = value }
-    }
-    // The position the player is *meant* to be at — the resume, last inbound jump, or last user
-    // navigation. The audio position may only DRIVE the ebook outbound once it has genuinely reached/
-    // advanced past this point; a transient book-start/pre-seek position below it must never win the
-    // cycle (the erase was a fresh-stamped 0 overwriting the read CFI). Below it, the cycle runs
-    // inbound-only so a newer remote (e.g. an ABS-web read) still pulls in. ADR 0035.
-    private var reconciledResumeSec: Double = 0.0
-
-    // Shared local-persistence policy with the ebook reader. Hot path (follow-loop tick): persist the
-    // listen position to the durable audiobook store so it survives process death and can win the
-    // last-update-wins resume. Cold path (pause/close): also write the `readingProgress` float to
-    // `library_items` so the detail/library screens reflect listening, without invalidating the library
-    // Room flow per tick. Backend (ABS) sync runs outside this coordinator (ADR 0035).
-    private val positionSaveCoordinator = com.riffle.app.feature.reader.PositionSaveCoordinator<Double>(
-        updateProgress = { progress -> updateReadingProgressUseCase(itemId, progress) },
-        savePosition = { pos ->
-            if (sourceId.isNotEmpty()) {
-                audiobookPositionStore.save(sourceId, itemId, pos)
-                // Matched book: listening is also reading — persist the translated reading position
-                // locally so the durable sweep pushes the ebook record too, without reopening (ADR 0036).
-                reconciliationCoordinator.mirrorListeningToReading(sourceId, itemId, pos)
-                // …and the readaloud resume, so reopening the reader and pressing Play lands on the
-                // listened sentence rather than a stale one (ADR 0037). Runs on every save (hot path
-                // and close) so a process-death mid-listen still leaves a fresh resume.
-                reconciliationCoordinator.writeListeningToReadaloud(sourceId, itemId, pos)
-            }
-        },
-    )
-
-    private val skipIntervalSec: StateFlow<Double> = listeningPreferencesStore.skipIntervalSeconds
-        .map { it.toDouble() }
-        .stateIn(viewModelScope, SharingStarted.Eagerly, ListeningPreferencesStore.DEFAULT_SKIP_INTERVAL_SECONDS.toDouble())
-
-    private val rewindIntervalSec: StateFlow<Double> = listeningPreferencesStore.rewindIntervalSeconds
-        .map { it.toDouble() }
-        .stateIn(viewModelScope, SharingStarted.Eagerly, ListeningPreferencesStore.DEFAULT_REWIND_INTERVAL_SECONDS.toDouble())
-
-    private val rewindOnResumeSec: StateFlow<Double> = listeningPreferencesStore.rewindOnResumeSeconds
-        .map { it.toDouble() }
-        .stateIn(viewModelScope, SharingStarted.Eagerly, ListeningPreferencesStore.DEFAULT_REWIND_ON_RESUME_SECONDS.toDouble())
-
-    private val _events = MutableSharedFlow<AudiobookPlayerEvent>(extraBufferCapacity = 1)
-    val events: SharedFlow<AudiobookPlayerEvent> = _events.asSharedFlow()
-
-    val uiState: StateFlow<AudiobookPlayerUiState> =
-        combine(meta, controller.state, controller.sleepTimer) { m, playback, timer ->
-            val pos = playback.positionSec
-            val chapter = timeline.chapterAt(pos)
-            m.copy(
-                isPlaying = playback.isPlaying,
-                speed = playback.speed,
-                positionSec = pos,
-                durationSec = if (playback.durationSec > 0) playback.durationSec else m.durationSec,
-                bufferedPositionSec = playback.bufferedSec,
-                currentChapterTitle = chapter?.title,
-                chapterStartsSec = timeline.chapters.map { it.startSec },
-                chapters = timeline.chapters,
-                currentChapterIndex = chapter?.index ?: -1,
-                canPreviousChapter = timeline.canPreviousChapter,
-                canNextChapter = timeline.canNextChapter,
-                bookmarks = m.bookmarks,
-                bookmarksOffline = m.bookmarksOffline,
-                sleepTimer = timer,
-            )
-        }.combine(skipIntervalSec) { state, skip ->
-            state.copy(skipIntervalSeconds = skip.toInt())
-        }.combine(rewindIntervalSec) { state, rewind ->
-            state.copy(rewindIntervalSeconds = rewind.toInt())
-        }.stateIn(viewModelScope, SharingStarted.Eagerly, AudiobookPlayerUiState(loading = true))
-
-    init {
-        viewModelScope.launch {
-            try {
-            val t0 = clock.nowMs()
-            logger.d(LogChannel.Handoff) { "AB.VM init start itemId=$itemId startAtSec=$startAtSec" }
-            val server = sourceRepository.getActive()
-            sourceId = server?.id ?: ""
-            // Claim this audiobook so the durable sweep leaves it to this player's own cycle (ADR 0036).
-            if (sourceId.isNotEmpty()) openReconcileTargets.markOpen(sourceId, itemId)
-            // Observe this book's bookmarks live into the UI state (ordered by position, earliest first).
-            if (sourceId.isNotEmpty()) {
-                viewModelScope.launch {
-                    bookmarkStore.observe(sourceId, itemId).collect { list ->
-                        meta.value = meta.value.copy(bookmarks = list)
-                    }
-                }
-                // Derive the quiet offline note: unsynced (dirty) bookmarks for this item AND offline.
-                // Sync is otherwise silent, so the note only shows when a push is genuinely blocked.
-                viewModelScope.launch {
-                    combine(
-                        bookmarkStore.observeHasUnsynced(sourceId, itemId),
-                        connectivityObserver.isOnline,
-                    ) { hasUnsynced, isOnline -> hasUnsynced && !isOnline }
-                        .collect { offline ->
-                            meta.value = meta.value.copy(bookmarksOffline = offline)
-                        }
-                }
-            }
-            val token = server?.let { tokenStorage.getToken(it.id) } ?: ""
-            logger.d(LogChannel.Handoff) { "AB.VM init: got server +${clock.nowMs() - t0}ms" }
-            val item = libraryObserver.getItem(itemId)
-            logger.d(LogChannel.Handoff) { "AB.VM init: got item +${clock.nowMs() - t0}ms" }
-            // Prefer a dedicated audiobook download, then a downloaded readaloud bundle's audio,
-            // then an auto-cached copy, then stream from ABS (connectivity-independent: a local copy
-            // always beats streaming). The streaming path launches a background job to cache all
-            // tracks and swap the player's unplayed queue items to file:// URLs on completion.
-            var launchCacheJob = false
-            var usedAutoCache = false
-            val session = if (sourceId.isEmpty()) null
-                else audiobookDownloadRepository.localSession(sourceId, itemId)
-                    ?.also { logger.d(LogChannel.Handoff) { "AB.VM init: local download session +${clock.nowMs() - t0}ms" } }
-                    ?: bundleAudiobookSource.localSession(sourceId, itemId)
-                    ?.also { logger.d(LogChannel.Handoff) { "AB.VM init: bundle session +${clock.nowMs() - t0}ms" } }
-                    ?: audiobookCacheRepository.localSession(sourceId, itemId)
-                    ?.also {
-                        usedAutoCache = true
-                        logger.d(LogChannel.Handoff) { "AB.VM init: auto-cache session +${clock.nowMs() - t0}ms" }
-                    }
-                    ?: audiobookRepository.openSession(sourceId, itemId)
-                    ?.also { launchCacheJob = true; logger.d(LogChannel.Handoff) { "AB.VM init: ABS network session +${clock.nowMs() - t0}ms" } }
-            if (item == null || session == null) {
-                logger.d(LogChannel.Handoff) { "AB.VM init: FAILED (item=${item != null} session=${session != null}) +${clock.nowMs() - t0}ms" }
-                meta.value = meta.value.copy(loading = false, failed = true)
-                return@launch
-            }
-            if (usedAutoCache) {
-                viewModelScope.launch { contentCacheAccessStore.markAccessed(audiobookCacheKey(sourceId, itemId)) }
-            }
-            timeline = session.timeline
-            // Resolve the resume position: last-update-wins reconcile (ADR 0035), progress-fraction
-            // fallback for offline-with-bundle-only, finished-book replay guard, and the readaloud→
-            // audiobook handoff override — all consolidated in AudiobookResumeResolver.
-            val resume = resumeResolver.resolve(
-                sourceId = sourceId,
-                itemId = itemId,
-                session = session,
-                readingProgressFraction = item.readingProgress,
-                startAtSec = startAtSec,
-            )
-            val resumeSec = resume.resumeSec
-            val resumeStamp = resume.resumeStamp
-            // Per-book speed (ADR 0034), shared with the linked Readaloud. Resolve the audio-settings
-            // key the *same* way the reader does — via the resolver on this audiobook's link — so both
-            // land on the identical key regardless of the `hasAudio` flag or sort order. With no link,
-            // settings key on this ABS item (an audiobook with no bundle gets its own row). Loaded
-            // before prepare so it's applied to the freshly-connected controller below; absent ⇒ 1×.
-            val link = readaloudLinkRepository.findByAbsItem(sourceId, itemId)
-            audioSettingsIdentity = if (link != null) {
-                audioIdentityResolver.resolveForStorytellerBook(link.storytellerSourceId, link.storytellerBookId)
-            } else {
-                AudioIdentity(sourceId, itemId)
-            }
-            val initialSpeed = audioPlaybackPreferencesStore.load(audioSettingsIdentity)
-                ?: listeningPreferencesStore.defaultPlaybackSpeed.first()
-            // Readaloud is only actually offerable when the synced bundle is present — the same gate the
-            // reader applies (readaloudControlState): a Storyteller book always qualifies, a matched ABS
-            // book only once its bundle is downloaded, an unmatched ABS book never. The bundle is keyed by
-            // the linked Storyteller book (or this item, on a Storyteller service), NOT the ABS item id.
-            val isStoryteller = server?.serverType == com.riffle.core.models.ServerType.STORYTELLER_SERVICE
-            val audioServerId = link?.storytellerSourceId ?: sourceId
-            val audioBookId = link?.storytellerBookId ?: itemId
-            val readaloudAvailable = com.riffle.app.feature.reader.readaloudControlState(
-                isStoryteller = isStoryteller,
-                isMatchedAbs = link != null,
-                bundlePresent = readaloudAudioRepository.isAudioAvailable(audioServerId, audioBookId),
-            ).enabled
-            // The readaloud EBOOK to switch to on swipe-down: among this Storyteller book's ABS targets,
-            // the readable one (the ebook in a split library); or this same item if it's a combined
-            // ebook+audio. Null when there's no readaloud ebook OR no synced bundle yet, which keeps the
-            // swipe-down hint hidden and the gesture inert until read-along can actually happen.
-            val readaloudEbookItemId: String? = if (!readaloudAvailable) null else link?.let { l ->
-                readaloudLinkRepository.findByStorytellerBook(l.storytellerSourceId, l.storytellerBookId)
-                    .firstOrNull { t ->
-                        t.absLibraryItemId != itemId &&
-                            libraryObserver.getItem(t.absSourceId, t.absLibraryItemId)?.isReadable == true
-                    }
-                    ?.absLibraryItemId
-                    ?: itemId.takeIf { item.isReadable }
-            }
-            // copy() (not a fresh state) so any bookmarks the live collector already observed during
-            // the suspend points above carry forward — a fresh state would wipe them, leaving an
-            // existing book's bookmarks stuck empty until Room's Flow next re-emits (on add/rename/delete).
-            meta.value = meta.value.copy(
-                loading = false,
-                failed = false,
-                title = item.title,
-                author = item.author,
-                publishedYear = item.publishedYear,
-                coverUrl = item.coverUrl,
-                authToken = token,
-                durationSec = session.timeline.durationSec,
-                readaloudEbookItemId = readaloudEbookItemId,
-                genres = item.genres,
-                facts = buildAudiobookFacts(session.timeline.durationSec, item.genres),
-                description = item.description,
-            )
-            // Rewind-on-resume also applies to reopening an in-progress book — the open path plays
-            // directly (below), bypassing togglePlayPause's resume rewind, so without this the setting
-            // would only ever fire on an in-player pause→play and look broken on the far more common
-            // "left and came back" resume. It shifts only where playback *starts*, NOT the resume floor
-            // (reconciledResumeSec stays at the true resumeSec below): seeding prepare() lower must not let
-            // the follow-loop / close-persist guards write that lower point back, or repeated open→close
-            // cycles would creep the saved position backward by rewindOnResume each time. Re-hearing the
-            // rewound seconds simply doesn't advance saved progress until playback passes the real resume.
-            // Read straight from the store (not the eager StateFlow, which may not have emitted this early in
-            // init). Guarded on resumeSec > 0 (a fresh/replayed book at the start has nothing to rewind) and
-            // on the normal-open branch only — the handoff must continue seamlessly from its hand-off point.
-            var playbackStartSec = resumeSec
-            if (startAtSec < 0.0) {
-                val rewindOnResume = listeningPreferencesStore.rewindOnResumeSeconds.first().toDouble()
-                if (rewindOnResume > 0.0 && resumeSec > 0.0) {
-                    playbackStartSec = (resumeSec - rewindOnResume).coerceAtLeast(0.0)
-                }
-            }
-
-            // Background auto-cache: download all tracks silently so the next open (or the current
-            // session after the swap) uses local files instead of the live stream. Launched before
-            // prepare() so the download starts as early as possible. viewModelScope cancels it if
-            // the user exits before caching completes; the partial dir is cleaned up by the impl.
-            if (launchCacheJob && session.timeline.durationSec > 0.0) {
-                val capturedSession = session
-                val capturedSourceId = sourceId
-                val capturedItemId = itemId
-                viewModelScope.launch {
-                    audiobookCacheRepository.awaitCachedAudiobook(capturedSourceId, capturedItemId, capturedSession)
-                    // Swap the unplayed tail of the queue to file:// URLs.
-                    val cachedSession = audiobookCacheRepository.localSession(capturedSourceId, capturedItemId) ?: return@launch
-                    contentCacheAccessStore.markAccessed(audiobookCacheKey(capturedSourceId, capturedItemId))
-                    val currentTrackIndex = com.riffle.core.models.AudiobookTracks.trackIndexAt(
-                        controller.currentAbsoluteSec(),
-                        capturedSession.tracks,
-                    )
-                    val nextIndex = currentTrackIndex + 1
-                    controller.swapTracksFromIndex(nextIndex, cachedSession.trackUrls.drop(nextIndex))
-                    logger.d(LogChannel.Handoff) { "AB.VM cache swap: tracks from $nextIndex swapped to local files" }
-                }
-            }
-
-            // Store resolved data so handoff activations (first swipe-up or re-activations after
-            // a swipe-down) can re-prepare the controller without re-fetching (ADR 0039).
-            resolvedSession = session
-            sessionDeferred.complete(session)
-            resolvedCoverUri = item.coverUrl
-            resolvedBookTitle = item.title
-            resolvedInitialSpeed = initialSpeed
-            logger.d(LogChannel.Handoff) { "AB.VM init: resolvedSession ready +${clock.nowMs() - t0}ms (startAtSec=$startAtSec)" }
-
-            if (startAtSec == PREWARM_SENTINEL) {
-                // Pre-connect the binder now so the first swipe-up pays ~0 ms instead of the full
-                // MediaController.Builder.buildAsync round-trip (ADR 0039).
-                logger.d(LogChannel.Handoff) { "AB.VM init: warming binder +${clock.nowMs() - t0}ms" }
-                controller.warmBinder()
-                logger.d(LogChannel.Handoff) { "AB.VM init: binder warm +${clock.nowMs() - t0}ms" }
-            } else {
-                // Normal open or readaloud→audiobook handoff via nav arg: prepare and play now.
-                // Resume at the server-recorded position (last-update-wins resume; ADR 0035).
-                controller.prepare(
-                    trackUrls = session.trackUrls,
-                    spans = session.tracks,
-                    durationSec = session.timeline.durationSec,
-                    startAtSec = playbackStartSec,
-                    localZipFile = session.localZipFile,
-                    coverUri = item.coverUrl,
-                    bookTitle = item.title,
-                    chapters = session.timeline.chapters,
-                )
-                // Record the active session so a media-notification tap reopens this audiobook player.
-                nowPlayingStore.set(com.riffle.app.playback.NowPlaying.Audiobook(itemId))
-                // Apply the persisted per-book speed to the freshly-prepared (singleton) controller,
-                // which would otherwise retain the previous book's speed. Set directly on the controller
-                // (not the VM's setSpeed) so restoring the saved value doesn't re-save it.
-                controller.setSpeed(initialSpeed)
-                reconciledResumeSec = resumeSec
-                localUpdatedAt = resumeStamp
-                // Opening the player is itself a "play" intent (the user tapped Listen), so start
-                // playback immediately rather than landing on a paused player. A genuinely-newer remote
-                // resume is still honoured: attachReaderSync's inbound-only reconcile seeks the
-                // already-playing position to the reconciled point (ADR 0035).
-                //
-                // Exception: if the book was detected as finished on open (resume near the end →
-                // reset to 0 by the finished-book guard), skip auto-play. This prevents the player
-                // from starting a "replay from the beginning" without user intent when Android
-                // restores the back stack after killing the process while the book was playing in
-                // the background. The user presses Play explicitly to re-listen.
-                if (!resume.wasFinishedOnOpen) controller.play()
-
-                // The cross-EPUB index build (needed for the full ebook-sync coordinator) is self-healed
-                // by ReaderSyncFactory.createIfApplicable below: if the bundle is present but the index
-                // isn't built yet, it enqueues the build there — so this open is itself the player-open
-                // retry path (ADR 0037). No explicit enqueue needed here.
-
-                // Attach the matched 2-peer cycle if its prerequisites are already cached (the follow
-                // loop re-attaches later if the index is still building). Seed it with the reconciled
-                // resume so a genuinely-newer local listen can lead the first cycle. Use resumeStamp
-                // (== resumeUpdatedAt normally, but the fresh now-stamp on a readaloud→audiobook
-                // handoff) so the handed-off position leads the first cycle.
-                attachReaderSync(resumeSec, resumeStamp)
-                // Always run the follow loop so the listen position reaches ABS while playing — in the
-                // matched case it drives both ABS peers (audio-led), otherwise it pushes the single ABS
-                // audiobook record (ADR 0035). Without this a plain audiobook only synced on pause/close.
-                followLoopOrchestrator.start(viewModelScope, followContext)
-            }
-            // PREWARM_SENTINEL: controller.prepare() is intentionally deferred. The binder is
-            // pre-connected above; the handoff watcher below receives the actual start position
-            // when the user swipes up.
-            } finally {
-                // Guarantee the deferred is always completed so activateFromHandoff() never
-                // hangs on await() when init fails or returns early (e.g. missing item/session).
-                if (!sessionDeferred.isCompleted) sessionDeferred.complete(null)
-            }
-        }
-
-        // Watch for the swipe-up handoff signal from EpubReaderViewModel. Runs in all modes so
-        // re-activations (second+ swipe-up after a swipe-down) also work.
-        viewModelScope.launch {
-            audiobookHandoffState.pendingHandoff
-                .filterNotNull()
-                .filter { it.itemId == itemId }
-                .collect { signal ->
-                    audiobookHandoffState.consumeHandoff()
-                    activateFromHandoff(signal.atSec)
-                }
-        }
-
-        // Watch for the overlay-dismissed signal (back button without a swipe-down handoff).
-        // Pause the controller so audiobook audio stops while the reader is visible.
-        viewModelScope.launch {
-            audiobookHandoffState.pendingDismiss
-                .filterNotNull()
-                .filter { it == itemId }
-                .collect {
-                    audiobookHandoffState.consumeDismiss()
-                    followLoopOrchestrator.cancel()
-                    controller.pause()
-                }
-        }
-
-        // End-of-book: when the player reports STATE_ENDED (last track played through), close the
-        // android-level player immediately and tell the screen to pop back to the detail view.
-        //
-        // We can't rely solely on onCleared() to tear the player down: by the time the screen pops
-        // and the VM is cleared, the player has been sitting in STATE_ENDED with media items still
-        // in the queue, and the [AudioPlayerService] foreground notification stays put for the user
-        // to see. Tear down here so the foreground notification drops right away; onCleared() runs
-        // the same teardown shortly after, idempotently (controller is already null on the second
-        // pass).
-        viewModelScope.launch {
-            controller.playbackEnded.collect {
-                // Guard against re-firing on the same VM. During auto-advance the destination
-                // VM briefly overlaps with the outgoing VM on the singleton controller — the
-                // listener's real-time tryEmit on STATE_ENDED can push another Unit to the
-                // SharedFlow buffer before the incoming VM's prepare() replaces media items,
-                // and every stray subscriber that consumes it would emit yet another
-                // PlaylistAdvance. Verified end-to-end: dropping this guard produced 3
-                // duplicate advances back-to-back within 1s in logcat.
-                if (handingOffToPlaylistAdvance) return@collect
-                flushPendingSpeed()
-                // Resolve next-in-playlist BEFORE deciding whether to tear the singleton player
-                // down. When auto-advancing, keep the MediaController connection alive so the
-                // incoming VM's controller.prepare() replaces media items on the SAME session —
-                // if we called pushProgressAndStopPlayer() here (or in onCleared), a race with
-                // VM2's init would release the connector after VM2 had already prepared+played,
-                // leaving the next book silently paused. Non-advance case (no playlist, or last
-                // item) keeps the original teardown so the foreground notification drops.
-                val nextInPlaylist = resolveNextPlaylistItem()
-                if (nextInPlaylist != null) {
-                    handingOffToPlaylistAdvance = true
-                    // Wipe the replay-cached STATE_ENDED Unit BEFORE the next VM subscribes,
-                    // or its collector re-consumes it and loop-advances through every remaining
-                    // item in milliseconds (verified end-to-end: two auto-advance log lines for
-                    // the same next-item id ~600ms apart, then playback dead).
-                    controller.clearEndOfBookCache()
-                    logger.d(LogChannel.Handoff) { "AB.VM playlist auto-advance → $nextInPlaylist" }
-                    _events.tryEmit(AudiobookPlayerEvent.PlaylistAdvance(nextInPlaylist))
-                } else {
-                    pushProgressAndStopPlayer()
-                    clearAudiobookNowPlaying()
-                    _events.tryEmit(AudiobookPlayerEvent.Finished)
-                }
-            }
-        }
-
-        // End-of-chapter sleep timer: fire when chapter index advances while EoC mode is active.
-        var eocPrevChapterIndex = -1
-        viewModelScope.launch {
-            uiState.collect { state ->
-                val idx = state.currentChapterIndex
-                if (eocPrevChapterIndex >= 0
-                    && idx > eocPrevChapterIndex
-                    && state.sleepTimer is SleepTimerMode.EndOfChapter
-                ) {
-                    controller.triggerSleepNow()
-                }
-                eocPrevChapterIndex = idx
-            }
-        }
-    }
-
-    private fun audiobookCacheKey(sourceId: String, itemId: String): ContentCacheKey =
-        ContentCacheKey(sourceId, itemId, ContentCacheArtifactKind.Audiobook)
-
-    /**
-     * Delegate to [reconciliationCoordinator], applying the returned jump (if any) to the
-     * controller + resume floor and adopting the canonical timestamp — wiring the coordinator's
-     * side-effect-free result into the VM's audio + floor state (ADR 0035).
-     */
-    private suspend fun attachReaderSync(atSec: Double, atUpdatedAt: Long): Boolean {
-        val result = reconciliationCoordinator.attach(sourceId, itemId, atSec, atUpdatedAt)
-        result.jumpToAudioSec?.let { controller.seekTo(it); reconciledResumeSec = it }
-        localUpdatedAt = maxOf(localUpdatedAt, result.canonicalLastUpdate)
-        return result.readerSyncAttached
-    }
-
-    fun togglePlayPause() {
-        if (controller.state.value.isPlaying) {
-            controller.pause()
-            followLoopOrchestrator.flushNow()
-        } else {
-            val rewindSec = rewindOnResumeSec.value
-            val posBeforeRewind = controller.currentAbsoluteSec()
-            if (rewindSec > 0) {
-                val newPos = (posBeforeRewind - rewindSec).coerceAtLeast(0.0)
-                reconciledResumeSec = newPos
-                controller.seekTo(newPos)
-            }
-            controller.play()
-        }
-    }
-
-    // A user seek is genuine navigation: move the resume baseline to the target so the new position
-    // (forward or backward) is allowed to drive the ebook, rather than being mistaken for a transient.
-    fun seekTo(positionSec: Double) {
-        reconciledResumeSec = positionSec
-        controller.seekTo(positionSec)
-    }
-
-    fun rewind() {
-        val rewindSec = rewindIntervalSec.value
-        reconciledResumeSec = (controller.currentAbsoluteSec() - rewindSec).coerceAtLeast(0.0)
-        controller.skipBy(-rewindSec)
-    }
-
-    fun forward() {
-        val skipSec = skipIntervalSec.value
-        reconciledResumeSec = controller.currentAbsoluteSec() + skipSec
-        controller.skipBy(skipSec)
-    }
-
-    fun previousChapter() {
-        timeline.previousChapterTargetSec(controller.currentAbsoluteSec())?.let { reconciledResumeSec = it; controller.seekTo(it) }
-    }
-
-    fun nextChapter() {
-        timeline.nextChapterTargetSec(controller.currentAbsoluteSec())?.let { reconciledResumeSec = it; controller.seekTo(it) }
-    }
-
-    /**
-     * The book-absolute listen position right now. The UI snapshots this when the New-bookmark dialog
-     * opens so the title, position label and the eventually-saved row all agree, even if playback keeps
-     * running while the user edits the title.
-     */
-    fun currentPositionSec(): Double = controller.currentAbsoluteSec()
-
-    /** The pre-filled (editable) default title for a new bookmark at [positionSec]. */
-    fun defaultBookmarkTitle(positionSec: Double): String =
-        BookmarkTitleBuilder.defaultTitle(timeline, positionSec)
-
-    /**
-     * Create a bookmark at [positionSec] (the position pinned when the dialog opened — NOT the live
-     * playhead, which may have drifted while the user typed). The new row arrives in
-     * [AudiobookPlayerUiState.bookmarks] via the store observation.
-     */
-    fun addBookmark(title: String, positionSec: Double) {
-        if (sourceId.isEmpty()) return
-        viewModelScope.launch { bookmarkStore.add(sourceId, itemId, positionSec, title, clock.nowMs()) }
-    }
-
-    fun renameBookmark(id: String, title: String) {
-        viewModelScope.launch { bookmarkStore.rename(id, title, clock.nowMs()) }
-    }
-
-    fun deleteBookmark(id: String) {
-        viewModelScope.launch { bookmarkStore.delete(id, clock.nowMs()) }
-    }
-
-    /** Jump the player to a saved bookmark's book-absolute position (genuine user navigation). */
-    fun seekToBookmark(positionSec: Double) {
-        seekTo(positionSec)
-    }
-
-    /**
-     * Set playback speed to a granular value (the shared speed sheet snaps to 0.05× steps). Applied to
-     * the live player immediately; persisted per book (ADR 0040) debounced so a scrub through many
-     * intermediate values writes only the settled speed, not every step.
-     */
-    fun setSpeed(speed: Float) {
-        controller.setSpeed(speed)
-        pendingSpeed = speed
-        speedSaveJob?.cancel()
-        speedSaveJob = viewModelScope.launch {
-            kotlinx.coroutines.delay(SPEED_SAVE_DEBOUNCE_MS)
-            if (sourceId.isEmpty()) return@launch
-            audioPlaybackPreferencesStore.save(audioSettingsIdentity, speed)
-            pendingSpeed = null
-        }
-    }
-
-    fun setSleepTimer(mode: SleepTimerMode) = controller.setSleepTimer(mode)
-    fun cancelSleepTimer() = controller.cancelSleepTimer()
-
-    /** Persist a speed change that the debounce window hadn't flushed yet (e.g. on close). */
-    private fun flushPendingSpeed() {
-        val speed = pendingSpeed ?: return
-        speedSaveJob?.cancel()
-        pendingSpeed = null
-        if (sourceId.isEmpty()) return
-        // progressFlushScope, not viewModelScope: onCleared cancels viewModelScope before calling
-        // this, so a launch there never executes (same reasoning as FollowLoopOrchestrator.stopWithFinalFlush).
-        progressFlushScope.flush { audioPlaybackPreferencesStore.save(audioSettingsIdentity, speed) }
-    }
-
-    // Set when the user swipes down to switch into readaloud. The audiobook and readaloud share ONE
-    // AudioPlayerService; readaloud takes over that player, so onCleared must NOT stop/clear it (that
-    // would pause the readaloud playback that just started). We save progress + release the handle
-    // here instead.
-    private var handingOffToReadaloud = false
-
-    // Set when this VM has emitted PlaylistAdvance and is about to be replaced by the next
-    // playlist item's VM on the same singleton controller. Same rationale as [handingOffToReadaloud]:
-    // onCleared MUST NOT release the shared connector, or a race with the incoming VM's prepare()
-    // silently kills its playback.
-    private var handingOffToPlaylistAdvance = false
-
-    /**
-     * Called when the user starts dragging down (before the threshold). Pre-resolves the SMIL seek
-     * target so [playFromSecond] in [ReadaloudController] can skip the computation at commit time
-     * (ADR 0039). No-op when no readaloud track is loaded (audiobook-only session).
-     */
-    fun hintReadaloudHandoff() {
-        readaloudController.preWarmSeek(controller.currentAbsoluteSec())
-    }
-
-    /** Discard the pre-warm if the drag was abandoned without crossing the threshold. */
-    fun cancelHandoffHint() {
-        readaloudController.cancelPreWarm()
-    }
-
-    /**
-     * Prepare to switch to the readaloud reader: persist the just-reached listen position (so the
-     * reader/readaloud resume is current) and release the audiobook's handle to the shared player
-     * WITHOUT stopping it, so readaloud can take it over and keep playing. Call right before navigating.
-     */
-    fun prepareReadaloudHandoff() {
-        if (handingOffToReadaloud) return
-        handingOffToReadaloud = true
-        // Cancel the follow loop now (VM stays alive in the always-mounted overlay, so onCleared
-        // won't do this until the user exits the reader entirely).
-        followLoopOrchestrator.stopWithFinalFlush()
-        controller.releaseForHandoff()
-    }
-
-    /**
-     * Prepare and start the audiobook controller from [atSec] on the concatenated timeline.
-     * Called by the handoff watcher when the user swipes up (for both first activation from
-     * pre-warm state and subsequent re-activations after a swipe-down).
-     */
-    private suspend fun activateFromHandoff(atSec: Double) {
-        val t0 = clock.nowMs()
-        logger.d(LogChannel.Handoff) { "AB.activateFromHandoff start atSec=$atSec resolvedSession=${resolvedSession != null}" }
-        val session = resolvedSession ?: run {
-            // Init still running — await the bundle-session fetch rather than drop the handoff.
-            logger.d(LogChannel.Handoff) { "AB.activateFromHandoff: init in progress, awaiting session (up to 10s)" }
-            withTimeoutOrNull(10_000L) { sessionDeferred.await() }
-        } ?: run {
-            logger.d(LogChannel.Handoff) { "AB.activateFromHandoff: DROPPED — session unavailable after waiting" }
-            return
-        }
-        handingOffToReadaloud = false
-        val finalSec = atSec.coerceIn(0.0, session.timeline.durationSec)
-        controller.prepare(
-            trackUrls = session.trackUrls,
-            spans = session.tracks,
-            durationSec = session.timeline.durationSec,
-            startAtSec = finalSec,
-            localZipFile = session.localZipFile,
-            coverUri = resolvedCoverUri,
-            bookTitle = resolvedBookTitle,
-            chapters = session.timeline.chapters,
-        )
-        logger.d(LogChannel.Handoff) { "AB.activateFromHandoff: prepare() done +${clock.nowMs() - t0}ms" }
-        controller.setSpeed(resolvedInitialSpeed)
-        reconciledResumeSec = finalSec
-        localUpdatedAt = clock.nowMs()
-        if (sourceId.isNotEmpty()) {
-            audiobookPositionStore.save(sourceId, itemId, finalSec)
-            audiobookPositionStore.updateLocalTimestamp(sourceId, itemId, localUpdatedAt)
-        }
-        nowPlayingStore.set(com.riffle.app.playback.NowPlaying.Audiobook(itemId))
-        controller.play()
-        logger.d(LogChannel.Handoff) { "AB.activateFromHandoff: play() called +${clock.nowMs() - t0}ms" }
-        attachReaderSync(finalSec, localUpdatedAt)
-        followLoopOrchestrator.start(viewModelScope, followContext)
-    }
-
-    /**
-     * Flush the just-ended position to ABS / the local position store, then tear down the
-     * MediaSession queue and release the [AudiobookController] (which dismisses the foreground
-     * notification). Idempotent: the controller's own stop() guards against a re-stop after release.
-     */
-    private fun pushProgressAndStopPlayer() {
-        followLoopOrchestrator.stopWithFinalFlush()
-        controller.stop()
-    }
-
-    /** Clear THIS book from the now-playing store so the mini-bar doesn't linger after the session ends. */
-    private fun clearAudiobookNowPlaying() {
-        nowPlayingStore.clearIf { it is com.riffle.app.playback.NowPlaying.Audiobook && it.itemId == itemId }
-    }
-
-    override fun onCleared() {
-        followLoopOrchestrator.cancel()
-        // Release this book to the durable sweep again (ADR 0036).
-        if (sourceId.isNotEmpty()) {
-            openReconcileTargets.markClosed(sourceId, itemId)
-            // Mirror the coordinator's markOpen: on the index-free fallback path readerSync stays null
-            // but audiobookFollow marked the ebook item open, so close via the same fallback chain —
-            // otherwise the ebook item leaks in the open set and the sweep skips it until process death.
-            reconciliationCoordinator.ebookItemIdForMarkClosed
-                ?.let { openReconcileTargets.markClosed(sourceId, it) }
-        }
-        flushPendingSpeed()
-        // On a readaloud handoff OR playlist auto-advance the shared session is being taken over
-        // by another surface (readaloud, or the next playlist item's VM). Stopping here would
-        // release the singleton connector out from under it.
-        if (!handingOffToReadaloud && !handingOffToPlaylistAdvance) pushProgressAndStopPlayer()
-        // Leaving the player stops playback (no mini-bar), so this session is no longer playing —
-        // except on auto-advance where the next item IS now the playing session; don't clear.
-        if (!handingOffToPlaylistAdvance) clearAudiobookNowPlaying()
-        // Flush dirty progress rows to WebDAV in-process (ADR 0063), mirroring ReconnectSyncKicker.
-        progressFlushScope.flush { runCatching { progressSweep.run() } }
-        super.onCleared()
-    }
-
-    /**
-     * When the player was opened inside a playlist ([playlistId] + [playlistLibraryId] set), find
-     * the item that comes AFTER the current [itemId] in that playlist. Returns null when: no
-     * playlist context, the playlist has vanished (auto-deleted or removed), current item isn't
-     * in the playlist (removed while playing), or current is the last item. Callers treat null as
-     * "no auto-advance — emit Finished as normal".
-     */
-    private suspend fun resolveNextPlaylistItem(): String? {
-        val pid = playlistId ?: return null
-        val lid = playlistLibraryId ?: return null
-        val playlist = runCatching { playlistsRepository.getPlaylist(lid, pid) }.getOrNull() ?: return null
-        val currentIndex = playlist.itemIds.indexOf(itemId).takeIf { it >= 0 } ?: return null
-        return playlist.itemIds.getOrNull(currentIndex + 1)
-    }
-
-    private companion object {
-        // Debounce window for persisting a speed change, so a granular scrub settles to a single write
-        // rather than one per intermediate 0.05× value (matches EpubReaderViewModel).
-        const val SPEED_SAVE_DEBOUNCE_MS = 400L
-        // Sentinel for startAtSec: the overlay is always mounted for pre-warming; the actual start
-        // position arrives via AudiobookHandoffState when the user swipes up.
-        const val PREWARM_SENTINEL = -2.0
-    }
-}
+// AudiobookPlayerViewModel and related UI types have moved to feature:player.
+typealias AudiobookPlayerViewModel = com.riffle.feature.player.AudiobookPlayerViewModel
+typealias AudiobookPlayerUiState = com.riffle.feature.player.AudiobookPlayerUiState
+typealias AudiobookPlayerEvent = com.riffle.feature.player.AudiobookPlayerEvent
