+++ app/src/test/kotlin/com/riffle/app/playback/NowPlayingStoreTest.kt
@@ -1,5 +1,6 @@
 package com.riffle.app.playback
 
+import com.riffle.feature.player.NowPlaying
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
