+++ core/data/src/androidMain/kotlin/com/riffle/core/data/CrossEpubIndexBuildTrigger.kt
@@ -1,14 +1,6 @@
 package com.riffle.core.data
 
-import com.riffle.core.models.ReadaloudLink
-
-/**
- * Narrow seam for scheduling a cross-EPUB index build for a matched book, so callers (the library
- * detail ViewModel's download-complete trigger, the reader-sync factory's on-open self-heal) depend
- * on the *intent* — not the I/O-heavy [CrossEpubIndexBuilderService] that fulfils it (ADR 0023/0031).
- * Keeping the surface this small also makes those callers trivially unit-testable.
- */
-interface CrossEpubIndexBuildTrigger {
-    /** Schedule an idempotent background build for [link]; returns immediately. */
-    fun enqueueBuild(link: ReadaloudLink)
-}
+// CrossEpubIndexBuildTrigger has moved to core:domain so feature-level KMP modules can depend on
+// it without pulling in Android-only core:data. Keeping this alias so existing app/core:data
+// imports compile unchanged.
+typealias CrossEpubIndexBuildTrigger = com.riffle.core.domain.CrossEpubIndexBuildTrigger
