+++ settings.gradle.kts
@@ -48,4 +48,5 @@ include(":core:catalog-komga")
 include(":feature:library")
 include(":feature:reader")
 include(":feature:source")
+include(":feature:player")
 include(":shared")
